requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
sentence-transformers==2.7.0
faiss-cpu==1.8.0
numpy==1.26.4
pydantic==2.7.4